Refactor resources page
