+++ features/apimgt/org.wso2.carbon.apimgt.publisher.feature/src/main/resources/publisher-new/.eslintignore
@@ -2,6 +2,6 @@
 /node_modules/**
 /coverage/**
 /source/src/app/components/Apis/Details/Security/**
-/source/src/app/components/Apis/Details/Resources/**
+/source/src/app/components/Apis/Details/Resources/old/**
 /source/src/app/components/Apis/Details/LifeCycle/**
 /source/src/app/components/Apis/Details/Documents/**
