+++ features/apimgt/org.wso2.carbon.apimgt.publisher.feature/src/main/resources/publisher-new/source/src/app/data/api.js
@@ -18,6 +18,7 @@
 import APIClientFactory from './APIClientFactory';
 import Utils from './Utils';
 import Resource from './Resource';
+import cloneDeep from 'lodash.clonedeep';
 
 /**
  * An abstract representation of an API
@@ -72,21 +73,21 @@ class API extends Resource {
      * which is API body friendly to use with REST api requests
      * Use this method instead of accessing the private _data object for
      * converting to a JSON representation of an API object.
-     * Note: This is shallow coping
+     * Note: This is deep coping, Use sparingly, Else will have a bad impact on performance
      * Basically this is the revers operation in constructor.
-     * This method simply iterate through all the object properties
-     * and copy their values to new object excluding the properties in excludes list.
-     * So use this method sparingly!!
+     * This method simply iterate through all the object properties (excluding the properties in `excludes` list)
+     * and copy their values to new object.
+     * So use this method with care!!
      * @memberof API
      * @param {Array} [userExcludes=[]] List of properties that are need to be excluded from the generated JSON object
      * @returns {JSON} JSON representation of the API
      */
-    toJSON(resource = this, userExcludes = []) {
+    toJSON(userExcludes = []) {
         var copy = {},
             excludes = ['_data', 'client', 'apiType', ...userExcludes];
-        for (var prop in resource) {
+        for (var prop in this) {
             if (!excludes.includes(prop)) {
-                copy[prop] = resource[prop];
+                copy[prop] = cloneDeep(this[prop]);
             }
         }
         return copy;
@@ -276,8 +277,7 @@ class API extends Resource {
         }
     }
 
-    save(query) {
-        query = query ? query : 'v2';
+    save(query = 'v2') {
         const promisedAPIResponse = this.client.then((client) => {
             const properties = client.spec.definitions.API.properties;
             const data = {};
@@ -860,7 +860,7 @@ class API extends Resource {
      * @param api {Object} Updated API object(JSON) which needs to be updated
      */
     update(updatedProperties) {
-        const updatedAPI = { ...this.toJSON(), ...this.toJSON(updatedProperties) };
+        const updatedAPI = { ...this.toJSON(), ...updatedProperties };
         const promisedUpdate = this.client.then(client => {
             const payload = {
                 apiId: updatedAPI.id,
